Update internals of DynamicGame.compute() to combine states/actions into "lines".
A line is a combination of the state and the actions agents take into
the state. Combining them into one makes map_blocks() easier, more efficient
and allows for some cleaner code in other places as well. It also fixes
a bug where, in fact, the code would not map_blocks() over actions_others, as
it should have.

To accomplish this, the following has been done:
- Make the Numpy array interface (.shape, .dtype, .__getitem__()
  tuple indexing) part of the InterpolatedFunction abc. (+tests)
- Implement this abc for ChebyshevInterpolatedFunction. (+tests)
- Change MultivariateInterpolatedFunction to expose this interface
  as well. Access the underlying value functions via the .vf attribute.
- Update the internals of DynamicGame. Instead of state, actions_players,
  actions_others, now only lines are passed around.