Remove unncessary dask.compute() from DynamicGame.compute()
This used to be necessary since optimal_actions() would be returned
from _inner_loop() as a Dask array. Now we parse the optimal actions
from lines just above the return, and the lines are already
compute()d before we do that. Hence, optimal_actions is already
calculated and the dask.compute() is a no-op.