Also update value functions when they converge.
In DynamicGame.calculate_value_function(), the ValueFunction object
would be update()d *after* checking for convergence. This means
that if the value function iteration converged, the value function
would not be updated in the last iteration and the value functions
as stored in value_functions would be off by a factor eps.