Fix bug where DynamicGame.calculate_value_function() would be non-deterministic.
The issue is that map_blocks() gave the same name to repeated
invocations. This lead to sometimes values for prev_value_function
being read from the wrong memory location (depending on timing),
so that the value function loop would stop prematurely, thinking
it had converged. Fix this by implementing __dask_tokenize__()
on ChebyshevInterpolatedFunction so that repeated invocations of
map_blocks() get different names if the value function gets updated.