Remove unneccessary checks for Sequence types in init_tests.
Some checks in int_tests.py would check if an attribute on DynamicGame,
e.g. n_actions, would be a Sequence if the corresponding parameter
passed to __init__() is a scalar. However, this is superfluous since
the next line always uses len() which will probably fail anyway.
Moreover, we do not actually want to impose that these elements are
Sequence's since ndarray is not a Sequence and we want the implementation
to be able to switch to ndarray's internally.