Change interface of DynamicGame.state_evolution() for derivatives
Makes the state evolution interface more flexible. It now handles
a number of possible states with associated probabilities. These
probabilities can be exact or potentially (later on) be used to
add a "simulation interface" to accommodate continuous distributions.
The interface also allows for derivatives of both the outcomes and
probabilities wrt the actions. Only works without inattention at
this moment, the inattention code is not quite correct in the current
form.