Make TransformedInterpolation a subtype of AbstractInterpolation.
Not only does this clean up the notation in some function definitions,
it also paves the way to introduce a new type that interpolates based
on policy function estimation (e.g. least squares).

To clear up some ambiguities with Interpolations.gradient() and such methods,
we have to (as a result) specify more precisely the arguments for the
overloaded methods for TransformedInterpolation.